requirements.txt
# Secrets
python-dotenv

# Embeddings
openai

# Vector-db
pinecone-client

# Q&A
langchain

# pdf reader
pypdf

# extra dependencies
cryptography

# extra dependencies
tiktoken